cuidando dos efeitos em cascata de criação, edição, e deleção para de project para com task, estilizando alerta de início de tarefa e regulando dinâmica do ritmo do alerta

+++ adagioapi/src/main/java/io/adagio/adagioapi/services/ProjectService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import io.adagio.adagioapi.dto.OperationType;
 import io.adagio.adagioapi.models.Project;
 import io.adagio.adagioapi.models.Task;
 import io.adagio.adagioapi.models.User;
@@ -20,18 +21,30 @@ public class ProjectService {
 	@Autowired
 	private ProjectRepository projectRepository;
 	
-	public long deleteTasksByProjectAndUser(Project project,User user) {
+	public long deleteTasksByProjectOrDesvinculateAndUser(Project project,User user,OperationType operation) {
 		long deletados = 0;
 		
 		List<Task> tasks = taskRepository.findByProjectAndUser(project,user);
 		
 		for(Task task: tasks) {
-			taskRepository.deleteByIdAndUser(task.getId(), user);
-			deletados++;
+			if(operation == OperationType.DELETE) {
+				taskRepository.deleteByIdAndUser(task.getId(), user);
+				deletados++;
+			} else if(operation == OperationType.EDIT) {
+				task.setProject(null);
+				taskRepository.save(task);
+			}
+
 		}
 		
-		System.out.println(deletados);
-		
 		return deletados;
+	}
+	
+	public void vinculateTasksToProject(List<Task> tasks, Project project) {
+		
+		for(Task task:tasks) {
+			task.setProject(project);
+			taskRepository.save(task);
+		}
 	}
 }
